start to split
