+++ crates/goose-providers/Cargo.toml
@@ -46,6 +46,8 @@ uuid = { workspace = true, features = ["v4", "std"] }
 async-trait = { workspace = true }
 strum = { workspace = true }
 tokio = { workspace = true }
+tokio-stream = { workspace = true, features = ["io-util"] }
+tokio-util = { workspace = true, features = ["compat"] }
 rand = { workspace = true }
 url = { workspace = true }
 pem = { version = "3.0.2", default-features = false, features = ["std"], optional = true }
