+++ crates/goose-providers/src/formats.rs
@@ -1,1 +1,2 @@
 pub mod openai;
+pub mod openai_responses;
