+++ crates/goose-providers/src/formats/openai_responses.rs
@@ -1,15 +1,15 @@
 use crate::conversation::message::{Message, MessageContent};
+use crate::conversation::token_usage::{ProviderUsage, Usage};
+use crate::errors::ProviderError;
+use crate::formats::openai::{
+    extract_reasoning_effort, is_openai_responses_model, openai_reasoning_effort_for_thinking,
+};
 use crate::mcp_utils::extract_text_from_resource;
+use crate::model::ModelConfig;
 use anyhow::Error;
 use async_stream::try_stream;
 use chrono;
 use futures::Stream;
-use goose_providers::conversation::token_usage::{ProviderUsage, Usage};
-use goose_providers::errors::ProviderError;
-use goose_providers::formats::openai::{
-    extract_reasoning_effort, is_openai_responses_model, openai_reasoning_effort_for_thinking,
-};
-use goose_providers::model::ModelConfig;
 use rmcp::model::{object, CallToolRequestParams, RawContent, Role, Tool};
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
@@ -945,8 +945,8 @@ where
 mod tests {
     use super::*;
     use crate::conversation::message::MessageContent;
+    use crate::model::ModelConfig;
     use futures::StreamExt;
-    use goose_providers::model::ModelConfig;
     use rmcp::model::CallToolRequestParams;
     use rmcp::object;
