+++ crates/goose-providers/src/http_status.rs
@@ -6,8 +6,8 @@
 
 use std::time::{Duration, SystemTime};
 
+use crate::errors::ProviderError;
 use chrono::{DateTime, NaiveDateTime, TimeZone, Utc};
-use goose_providers::errors::ProviderError;
 use reqwest::header::{HeaderMap, RETRY_AFTER};
 use reqwest::{Response, StatusCode};
 use serde_json::Value;
