+++ crates/goose-providers/src/lib.rs
@@ -5,10 +5,14 @@ pub mod conversation;
 pub mod errors;
 pub mod formats;
 pub mod goose_mode;
+pub mod http_status;
 pub mod images;
 pub mod json;
 pub(crate) mod mcp_utils;
 pub mod model;
+pub mod openai;
+pub mod openai_compatible;
+pub mod openai_def;
 pub mod permission;
 pub mod request_log;
 pub mod retry;
