+++ crates/goose-providers/src/openai_compatible.rs
@@ -1,8 +1,8 @@
+use crate::conversation::token_usage::ProviderUsage;
+use crate::images::ImageFormat;
 use anyhow::Error;
 use async_stream::try_stream;
 use futures::TryStreamExt;
-use goose_providers::conversation::token_usage::ProviderUsage;
-use goose_providers::images::ImageFormat;
 use reqwest::Response;
 #[cfg(test)]
 use reqwest::StatusCode;
@@ -16,13 +16,13 @@ use super::api_client::ApiClient;
 use super::base::{stream_from_single_message, MessageStream, Provider};
 use super::retry::ProviderRetry;
 use crate::conversation::message::Message;
-use crate::providers::formats::openai_responses::responses_api_to_streaming_message;
-use goose_providers::errors::ProviderError;
-use goose_providers::formats::openai::{
+use crate::errors::ProviderError;
+use crate::formats::openai::{
     create_request, get_usage, response_to_message, response_to_streaming_message,
 };
-use goose_providers::model::ModelConfig;
-use goose_providers::request_log::{start_log, LoggerHandleExt, RequestLogHandle};
+use crate::formats::openai_responses::responses_api_to_streaming_message;
+use crate::model::ModelConfig;
+use crate::request_log::{start_log, LoggerHandleExt, RequestLogHandle};
 use rmcp::model::Tool;
 
 pub struct OpenAiCompatibleProvider {
@@ -228,7 +228,7 @@ pub fn stream_responses_compat(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use goose_providers::model::ModelConfig;
+    use crate::model::ModelConfig;
     use serde_json::json;
     use test_case::test_case;
